+++ public_html/install/index.php
@@ -54,7 +54,7 @@ if($stream)
     // Enable Autoloading with doctrine
     $classLoader = new ClassLoader('install'   ,   DOC_ROOT);
     $classLoader->register();
-    $classLoader = new ClassLoader('app'   ,   DOC_ROOT);
+    $classLoader = new ClassLoader('src'   ,   DOC_ROOT);
     $classLoader->register();
     $classLoader = null;
