+++ public_html/src/Business/Logic/PasswordHasher.php
@@ -1,6 +1,6 @@
 <?PHP
 
-namespace app\config;
+namespace src\Business\Logic;
 
 class PasswordHasher
 {
